backend/requirements.txt
fastapi==0.111.0
uvicorn==0.23.2
sqlalchemy==2.0.23
psycopg[binary]
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==3.2.0 # Downgraded for compatibility with passlib
python-jose[cryptography]==3.3.0
pyjwt==2.8.0
pydantic==2.7.1
cohere
qdrant-client==1.7.0
markdown